+++ kitty/options/to-c-generated.h
@@ -616,6 +616,19 @@ convert_from_opts_detect_urls(PyObject *py_opts, Options *opts) {
 }
 
 static void
+convert_from_python_detect_url_regex(PyObject *val, Options *opts) {
+    detect_url_regex(val, opts);
+}
+
+static void
+convert_from_opts_detect_url_regex(PyObject *py_opts, Options *opts) {
+    PyObject *ret = PyObject_GetAttrString(py_opts, "detect_url_regex");
+    if (ret == NULL) return;
+    convert_from_python_detect_url_regex(ret, opts);
+    Py_DECREF(ret);
+}
+
+static void
 convert_from_python_url_excluded_characters(PyObject *val, Options *opts) {
     url_excluded_characters(val, opts);
 }
@@ -1659,6 +1672,8 @@ convert_opts_from_python_opts(PyObject *py_opts, Options *opts) {
     convert_from_opts_url_prefixes(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_detect_urls(py_opts, opts);
+    if (PyErr_Occurred()) return false;
+    convert_from_opts_detect_url_regex(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_url_excluded_characters(py_opts, opts);
     if (PyErr_Occurred()) return false;